pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gdoc2netcfg"
version = "0.1.0"
description = "Generate network infrastructure config files from Google Spreadsheet data"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "cryptography",
    "google-auth-oauthlib>=1.0",
    "gspread>=6.0",
    "jinja2",
    "paho-mqtt>=2",
    "pysnmp>=7.0",
    "websockets>=13.0",
]

[project.scripts]
gdoc2netcfg = "gdoc2netcfg.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gdoc2netcfg", "src/nsdp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]
